Create a new ruby-graphviz file to demonstrate image generation
