+++ ruby/binary_search_trees/binary_search_trees.rb
@@ -1,14 +1,5 @@
 # THESE METHODS RESEMBLE TEXTBOOK PSEUDO CODE OR PROFESSOR CALLAHAN'S PYTHON CODE
 
-# INDEX
-#      inorder_tree_walk
-#      preoder_tree_walk
-#      postorder_tree_walk
-#      tree_search
-#      iterative_tree_search
-#      tree_minimum
-#      tree_maximum
-#      tree_successor
 #      tree_insert
 #      tree_delete -> transplant
 #
@@ -67,51 +58,6 @@ end
 #              /  \    /
 #             C   E   H
 #   tree = Tree.new(F)   NOTE: Whole tree is assumed to have been constructed alread
-
-# TODO: Add documentation on how to install GRAPHVIZ AND RUBY-GRAPHVIZ GEM
-# Public: Same as tree_insert method only here this generates a graphical image(.PNG)
-#         for tree visualization. Added a new method to not mix the actual logic with
-#         this gem specific code
-#
-# NOTE: Images are generated at a location where the code is executed
-#       Eg: If you are at /Users/whoami/Documents directory and ran the code
-#           by executing ruby /algorithms/ruby/binary_trees/binary_tree.rb, then
-#           images will be generated at /Users/whoami/Documents location and not
-#           at the location of the ruby file
-#
-# t - Tree structure
-# z - Node to be INSERTED
-# g - GraphViz instance to draw and update the graph
-def tree_insert_graphical(t, z, g)
-  return if (t.nil? || z.nil? || g.nil?)
-  y = nil
-  x = t.root
-  while !x.nil?
-    y = x
-    (z.key < x.key) ? (x = x.left) : (x = x.right)
-  end
-  z.p = y
-
-  # TODO: Document this
-  if y.nil?
-    t.root = z
-    g.add_nodes( "#{z.key}", "shape" => "record", "label" => "<left> left|<key> #{z.key}|<p> p|<right> right" )
-  elsif z.key < y.key
-    y.left = z
-    g.add_nodes( "#{y.key}", "shape" => "record", "label" => "<left> left|<key> #{y.key}|<p> parent|<right> right" )
-    g.add_nodes( "#{z.key}", "shape" => "record", "label" => "<left> left|<key> #{z.key}|<p> parent|<right> right" )
-
-    g.add_edges( { "#{y.key}" => :left}, {"#{z.key}" => :key} )
-    g.add_edges( { "#{z.key}" => :p}, {"#{y.key}" => :key} )
-  else
-    y.right = z
-    g.add_nodes( "#{y.key}", "shape" => "record", "label" => "<left> left|<key> #{y.key}|<p> parent|<right> right" )
-    g.add_nodes( "#{z.key}", "shape" => "record", "label" => "<left> left|<key> #{z.key}|<p> parent|<right> right" )
-
-    g.add_edges( { "#{y.key}" => :right}, {"#{z.key}" => :key} )
-    g.add_edges( { "#{z.key}" => :p}, {"#{y.key}" => :key} )
-  end
-end
 
 # TEST inorder_tree_walk
 #      preoder_tree_walk
